Test XSS bugs identified by ITHC
 - moved the tests to be next to the others for the same page
 - more assertions per code review
 - remove dependency on urllib
