+++ tests/app/main/views/test_buyers.py
@@ -21,6 +21,19 @@ class TestBuyersView(LoggedInApplicationTest):
 
         self.assertEqual("There are no opportunities with ID 1", banner_message)
 
+    def test_brief_not_found_flash_message_injection(self, data_api_client):
+        """
+        Asserts that raw HTML in a bad brief ID cannot be injected into a flash message.
+        """
+        # impl copied from test_should_display_a_useful_message_if_no_brief_found
+        data_api_client.get_brief.return_value = None
+        response = self.client.get('admin/buyers?brief_id=1%3Cimg%20src%3Da%20onerror%3Dalert%281%29%3E')
+        self.assertEqual(response.status_code, 404)
+
+        html_response = response.get_data(as_text=True)
+        self.assertNotIn("1<img src=a onerror=alert(1)>", html_response)
+        self.assertIn("1&lt;img src=a onerror=alert(1)&gt;", html_response)
+
     def test_table_should_show_a_useful_message_if_no_users(self, data_api_client):
         data_api_client.get_brief.return_value = {
             'briefs': {
